pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
authors = ["leeparnell <152523161+leeparnell@users.noreply.github.com>"]
description = ""
name = "pi-haiku"
packages = [{from = "src", include = "pi_haiku"}]
readme = "README.md"
version = "0.2.2"

[tool.poetry.dependencies]
python = "^3.11"
gitpython = "^3.1.43"
pi-conf = "^0.8.5.1"
pydantic-settings = "^2.3.4"
pygithub = "^2.3.0"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
toml = "^0.10.2"
toml-sort = "^0.23.1"
types-requests = "^2.32.0.20240712"
types-toml = "^0.10.8.20240310"

[tool.poetry.scripts]
haiku = "cli.haiku_cli:main"

[tool.tomlsort]
all = true
in_place = true
overrides."tool.poetry.dependencies".first = ["python"]
